Moved new Profile UI to UserProfile component
